Move DNS recovery code to own files
