+++ client/internal/dns/host_android.go
@@ -24,7 +24,3 @@ func (a androidHostManager) supportCustomPort() bool {
 func (a androidHostManager) restoreUncleanShutdownDNS(*netip.Addr) error {
 	return nil
 }
-
-func CheckUncleanShutdown(string) error {
-	return nil
-}
