+++ client/internal/dns/host_ios.go
@@ -41,7 +41,3 @@ func (a iosHostManager) supportCustomPort() bool {
 func (a iosHostManager) restoreUncleanShutdownDNS(*netip.Addr) error {
 	return nil
 }
-
-func CheckUncleanShutdown(string) error {
-	return nil
-}
