+++ client/internal/dns/unclean_shutdown_ios.go
@@ -1,0 +1,5 @@
+package dns
+
+func CheckUncleanShutdown(string) error {
+	return nil
+}
